second version of AnimalTree with constructor overloading; statics; more wide abstract classes usage.
