+++ src/training/oop/Pigeon.java
@@ -1,15 +1,15 @@
-package training.codeReuse;
+package training.oop;
 
 /**
  * Created by Ivan_Stepantsov on 8/3/2017.
  */
-class Swordfish extends AquaticAnimal {
+class Pigeon extends OrnithoAnimal {
     public void birth() {
-        System.out.println("Birth of fish. swim!");
+        System.out.println("Birth of bird. Fly!");
     }
 
     public void nutrition(String nutrition) {
-        System.out.format("Nutrition of organism made mostly by other smaller fish and %s.\n", nutrition);
+        System.out.format("Nutrition of Organism made mostly by insects and %s\n", nutrition);
     }
 
     @Override
