+++ src/training/oop/Swordfish.java
@@ -1,15 +1,15 @@
-package training.codeReuse;
+package training.oop;
 
 /**
  * Created by Ivan_Stepantsov on 8/3/2017.
  */
-class Pigeon extends OrnithoAnimal {
+class Swordfish extends AquaticAnimal {
     public void birth() {
-        System.out.println("Birth of bird. Fly!");
+        System.out.println("Birth of fish. swim!");
     }
 
     public void nutrition(String nutrition) {
-        System.out.format("Nutrition of Organism made mostly by insects and %s\n", nutrition);
+        System.out.format("Nutrition of organism made mostly by other smaller fish and %s.\n", nutrition);
     }
 
     @Override
